# ePurchase-service/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "ePurchase-API's",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "summary": "Get All Users",
        "tags": [
          "Users"
        ],
        "operationId": "GetAllUsers",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "post": {
        "summary": "Add user",
        "tags": [
          "Users"
        ],
        "operationId": "Adduser",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/AdduserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/users/8": {
      "get": {
        "summary": "Get User by Id",
        "tags": [
          "Users"
        ],
        "operationId": "GetUserbyId",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      },
      "delete": {
        "summary": "Delete User by Id",
        "tags": [
          "Users"
        ],
        "operationId": "DeleteUserbyId",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/users/9": {
      "put": {
        "summary": "Update user by id",
        "tags": [
          "Users"
        ],
        "operationId": "Updateuserbyid",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UpdateuserbyidRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/setupdb": {
      "get": {
        "summary": "Set up DB",
        "tags": [
          "Misc"
        ],
        "operationId": "SetupDB",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    }
  },
  "definitions": {
    "UpdateuserbyidRequest": {
      "title": "UpdateuserbyidRequest",
      "example": {
        "firstName": "shivam100",
        "cartDetails": [
          {
            "abcd": 253,
            "klmn": 5
          }
        ]
      },
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "cartDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartDetail"
          }
        }
      },
      "required": [
        "firstName",
        "cartDetails"
      ]
    },
    "CartDetail": {
      "title": "CartDetail",
      "example": {
        "abcd": 253,
        "klmn": 5
      },
      "type": "object",
      "properties": {
        "abcd": {
          "type": "integer",
          "format": "int32"
        },
        "klmn": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "abcd",
        "klmn"
      ]
    },
    "AdduserRequest": {
      "title": "AdduserRequest",
      "example": {
        "firstName": "shivam",
        "lastName": "singh",
        "email": "klmn1@gmail.com",
        "dob": "02-09-1998",
        "phone": "2563145873",
        "cartDetails": [
          {
            "klmn": 5,
            "ijkl": 253
          },
          {
            "klmn": 22,
            "ijkl": 123
          }
        ],
        "address": "Kanpur",
        "password": "password"
      },
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "dob": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "cartDetails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartDetail1"
          }
        },
        "address": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "firstName",
        "lastName",
        "email",
        "dob",
        "phone",
        "cartDetails",
        "address",
        "password"
      ]
    },
    "CartDetail1": {
      "title": "CartDetail1",
      "example": {
        "klmn": 5,
        "ijkl": 253
      },
      "type": "object",
      "properties": {
        "klmn": {
          "type": "integer",
          "format": "int32"
        },
        "ijkl": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "klmn",
        "ijkl"
      ]
    }
  },
  "tags": [
    {
      "name": "Users"
    },
    {
      "name": "Misc",
      "description": ""
    }
  ]
}